requirements.txt
requests
gtts
google-generativeai
